明细跳转功能

+++ app/src/main/java/com/keyou/fdcda/app/dao/ImageInfoMapper.java
@@ -1,14 +1,12 @@
 package com.keyou.fdcda.app.dao;
 
-import java.math.BigDecimal;
-
 import org.springframework.stereotype.Repository;
 
 import com.keyou.fdcda.api.model.ImageInfo;
 import com.keyou.fdcda.app.dao.base.BaseMapper;
 
 @Repository
-public interface ImageInfoMapper extends BaseMapper<ImageInfo, BigDecimal> {	
+public interface ImageInfoMapper extends BaseMapper<ImageInfo, Long> {	
 
 	/**
 	 * 调用存储过程
